requirements.txt
# Core Computer Vision e AI
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0

# Image Processing
pillow>=10.0.0

# Virtual Camera
pyvirtualcam>=0.10.0

# GPU Acceleration (Opzionale)
pyopencl>=2023.1.0

# Development Tools (Opzionali)
pytest>=7.0.0